requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
bcrypt>=4.0,<5.0
PyJWT>=2.8,<3.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

